feat: 대시보드 정보 연결
